Unify separated IrElementTransformer into a single implementation
