+++ backintime-plugin/compiler/src/main/kotlin/com/github/kitakkun/backintime/compiler/backend/BackInTimeIrGenerationExtension.kt
@@ -1,6 +1,6 @@
 package com.github.kitakkun.backintime.compiler.backend
 
-import com.github.kitakkun.backintime.compiler.backend.transformer.DebuggableStateHolderTransformer
+import com.github.kitakkun.backintime.compiler.backend.transformer.BackInTimeDebuggableClassTransformer
 import com.github.kitakkun.backintime.compiler.configuration.BackInTimeCompilerConfiguration
 import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
@@ -13,7 +13,7 @@ class BackInTimeIrGenerationExtension(
     override fun generate(moduleFragment: IrModuleFragment, pluginContext: IrPluginContext) {
         val context = BackInTimePluginContext(baseContext = pluginContext, config = config, moduleFragment = moduleFragment)
         with(context) {
-            moduleFragment.transformChildrenVoid(DebuggableStateHolderTransformer())
+            moduleFragment.transformChildrenVoid(BackInTimeDebuggableClassTransformer())
         }
     }
 }
